HTBHF-2037 Moved address features into address folder.
