+++ src/web/routes/application/address/constants.js
@@ -1,0 +1,14 @@
+module.exports.OS_PLACES_ADDRESS_KEYS = [
+  'ADDRESS',
+  'ORGANISATION_NAME',
+  'SUB_BUILDING_NAME',
+  'BUILDING_NAME',
+  'BUILDING_NUMBER',
+  'DEPENDENT_THOROUGHFARE_NAME',
+  'THOROUGHFARE_NAME',
+  'DOUBLE_DEPENDENT_LOCALITY',
+  'DEPENDENT_LOCALITY',
+  'POST_TOWN',
+  'POSTCODE',
+  'LOCAL_CUSTODIAN_CODE_DESCRIPTION'
+]
