+++ src/web/routes/application/common/constants.js
@@ -9,17 +9,3 @@ module.exports.EMAIL = 'email'
 module.exports.TEXT = 'text'
 module.exports.CONFIRMATION_CODE_SESSION_PROPERTY = 'confirmationCode'
 module.exports.CONFIRMATION_CODE_ENTERED_SESSION_PROPERTY = 'confirmationCodeEntered'
-module.exports.OS_PLACES_ADDRESS_KEYS = [
-  'ADDRESS',
-  'ORGANISATION_NAME',
-  'SUB_BUILDING_NAME',
-  'BUILDING_NAME',
-  'BUILDING_NUMBER',
-  'DEPENDENT_THOROUGHFARE_NAME',
-  'THOROUGHFARE_NAME',
-  'DOUBLE_DEPENDENT_LOCALITY',
-  'DEPENDENT_LOCALITY',
-  'POST_TOWN',
-  'POSTCODE',
-  'LOCAL_CUSTODIAN_CODE_DESCRIPTION'
-]
