+++ src/web/routes/application/steps.js
@@ -2,7 +2,7 @@ const { nationalInsuranceNumber } = require('./national-insurance-number')
 const { name } = require('./name')
 const { dateOfBirth } = require('./date-of-birth')
 const { areYouPregnant } = require('./are-you-pregnant')
-const { manualAddress } = require('./manual-address')
+const { manualAddress } = require('./address/manual-address')
 const { phoneNumber } = require('./phone-number')
 const { scotland } = require('./scotland')
 const { inScotland } = require('./in-scotland')
@@ -11,8 +11,8 @@ const { doYouHaveChildren } = require('./do-you-have-children')
 const { childDateOfBirth } = require('./child-date-of-birth')
 const { sendCode } = require('./send-code')
 const { enterCode } = require('./enter-code')
-const { postcode } = require('./postcode')
-const { selectAddress } = require('./select-address')
+const { postcode } = require('./address/postcode')
+const { selectAddress } = require('./address/select-address')
 
 const steps = [
   scotland,
